requirements.txt
langchain
openai
kor
tiktoken
tqdm
python-dotenv